Try moving the stacked tests to the UI test suite
